+++ services/ifrs/src/ifrs/core/file_registry.py
@@ -19,18 +19,12 @@ import logging
 import uuid
 from contextlib import suppress
 
-from ghga_event_schemas import pydantic_ as event_schemas
 from ghga_service_commons.utils.multinode_storage import ObjectStorages
 
 from ifrs.config import Config
 from ifrs.core import models
-from ifrs.core.utils import assert_record_is_new, make_record_from_update
 from ifrs.ports.inbound.file_registry import FileRegistryPort
-from ifrs.ports.outbound.dao import (
-    FileMetadataDaoPort,
-    OutboxDaoCollectionPort,
-    ResourceNotFoundError,
-)
+from ifrs.ports.outbound.dao import FileMetadataDaoPort, ResourceNotFoundError
 from ifrs.ports.outbound.event_pub import EventPublisherPort
 
 log = logging.getLogger(__name__)
@@ -39,11 +33,10 @@ log = logging.getLogger(__name__)
 class FileRegistry(FileRegistryPort):
     """A service that manages a registry files stored on a permanent object storage."""
 
-    def __init__(  # noqa: PLR0913
+    def __init__(
         self,
         *,
         file_metadata_dao: FileMetadataDaoPort,
-        outbox_dao_collection: OutboxDaoCollectionPort,
         event_publisher: EventPublisherPort,
         object_storages: ObjectStorages,
         config: Config,
@@ -51,7 +44,6 @@ class FileRegistry(FileRegistryPort):
         """Initialize with essential config params and outbound adapters."""
         self._event_publisher = event_publisher
         self._file_metadata_dao = file_metadata_dao
-        self._outbox_dao_collection = outbox_dao_collection
         self._object_storages = object_storages
         self._config = config
 
@@ -304,90 +296,3 @@ class FileRegistry(FileRegistryPort):
             "Finished object storage and metadata deletion for file ID '%s'", file_id
         )
         await self._event_publisher.file_deleted(file_id=file_id)
-
-    async def upsert_nonstaged_file_requested(
-        self, *, resource_id: str, update: event_schemas.NonStagedFileRequested
-    ) -> None:
-        """Upsert a NonStagedFileRequested event. Call `stage_registered_file` if the
-        idempotence check is passed.
-
-        Args:
-            resource_id:
-                The resource ID.
-            update:
-                The NonStagedFileRequested event to upsert.
-        """
-        dao = self._outbox_dao_collection.get_nonstaged_file_requested_dao()
-
-        record = make_record_from_update(models.NonStagedFileRequestedRecord, update)
-        if await assert_record_is_new(
-            dao=dao, resource_id=resource_id, update=update, record=record
-        ):
-            await self.stage_registered_file(
-                file_id=resource_id,
-                decrypted_sha256=update.decrypted_sha256,
-                outbox_object_id=update.target_object_id,
-                outbox_bucket_id=update.target_bucket_id,
-            )
-            await dao.insert(record)
-
-    async def upsert_file_deletion_requested(
-        self, *, resource_id: str, update: event_schemas.FileDeletionRequested
-    ) -> None:
-        """Upsert a FileDeletionRequested event. Call `delete_file` if the idempotence
-        check is passed.
-
-        Args:
-            resource_id:
-                The resource ID.
-            update:
-                The FileDeletionRequested event to upsert.
-        """
-        dao = self._outbox_dao_collection.get_file_deletion_requested_dao()
-
-        record = make_record_from_update(models.FileDeletionRequestedRecord, update)
-        if await assert_record_is_new(
-            dao=dao, resource_id=resource_id, update=update, record=record
-        ):
-            await self.delete_file(file_id=resource_id)
-            await dao.insert(record)
-
-    async def upsert_file_upload_validation_success(
-        self, *, resource_id: str, update: event_schemas.FileUploadValidationSuccess
-    ) -> None:
-        """Upsert a FileUploadValidationSuccess event. Call `register_file` if the
-        idempotence check is passed.
-
-        Args:
-            resource_id:
-                The resource ID.
-            update:
-                The FileUploadValidationSuccess event to upsert.
-        """
-        dao = self._outbox_dao_collection.get_file_upload_validation_success_dao()
-
-        record = make_record_from_update(
-            models.FileUploadValidationSuccessRecord, update
-        )
-        if await assert_record_is_new(
-            dao=dao, resource_id=resource_id, update=update, record=record
-        ):
-            file_without_object_id = models.FileMetadataBase(
-                file_id=update.file_id,
-                decrypted_sha256=update.decrypted_sha256,
-                decrypted_size=update.decrypted_size,
-                upload_date=update.upload_date,
-                decryption_secret_id=update.decryption_secret_id,
-                encrypted_part_size=update.encrypted_part_size,
-                encrypted_parts_md5=update.encrypted_parts_md5,
-                encrypted_parts_sha256=update.encrypted_parts_sha256,
-                content_offset=update.content_offset,
-                storage_alias=update.s3_endpoint_alias,
-            )
-
-            await self.register_file(
-                file_without_object_id=file_without_object_id,
-                staging_object_id=update.object_id,
-                staging_bucket_id=update.bucket_id,
-            )
-            await dao.insert(record)
