+++ services/ifrs/src/ifrs/inject.py
@@ -29,11 +29,13 @@ from ifrs.adapters.inbound.event_sub import (
     FileValidationSuccessListener,
     NonstagedFileRequestedListener,
 )
+from ifrs.adapters.inbound.idempotent import IdempotenceHandler
 from ifrs.adapters.outbound import dao
 from ifrs.adapters.outbound.event_pub import EventPubTranslator
 from ifrs.config import Config
 from ifrs.core.file_registry import FileRegistry
 from ifrs.ports.inbound.file_registry import FileRegistryPort
+from ifrs.ports.inbound.idempotent import IdempotenceHandlerPort
 
 
 @asynccontextmanager
@@ -42,7 +44,6 @@ async def prepare_core(*, config: Config) -> AsyncGenerator[FileRegistryPort, No
     dao_factory = MongoDbDaoFactory(config=config)
     object_storages = S3ObjectStorages(config=config)
     file_metadata_dao = await dao.get_file_metadata_dao(dao_factory=dao_factory)
-    outbox_dao_collection = await dao.get_outbox_dao_collection(dao_factory=dao_factory)
 
     async with KafkaEventPublisher.construct(config=config) as kafka_event_publisher:
         event_publisher = EventPubTranslator(
@@ -50,7 +51,6 @@ async def prepare_core(*, config: Config) -> AsyncGenerator[FileRegistryPort, No
         )
         file_registry = FileRegistry(
             file_metadata_dao=file_metadata_dao,
-            outbox_dao_collection=outbox_dao_collection,
             event_publisher=event_publisher,
             object_storages=object_storages,
             config=config,
@@ -73,7 +73,10 @@ def prepare_core_with_override(
 
 @asynccontextmanager
 async def prepare_outbox_subscriber(
-    *, config: Config, core_override: Optional[FileRegistryPort] = None
+    *,
+    config: Config,
+    core_override: Optional[FileRegistryPort] = None,
+    idempotence_handler_override: Optional[IdempotenceHandlerPort] = None,
 ) -> AsyncGenerator[KafkaOutboxSubscriber, None]:
     """Construct and initialize an event subscriber with all its dependencies.
     By default, the core dependencies are automatically prepared but you can also
@@ -82,10 +85,34 @@ async def prepare_outbox_subscriber(
     async with prepare_core_with_override(
         config=config, core_override=core_override
     ) as file_registry:
+        idempotence_handler = idempotence_handler_override
+        if not idempotence_handler:
+            dao_factory = MongoDbDaoFactory(config=config)
+            nonstaged_file_requested_dao = await dao.get_nonstaged_file_requested_dao(
+                dao_factory=dao_factory
+            )
+            file_upload_validation_success_dao = (
+                await dao.get_file_upload_validation_success_dao(
+                    dao_factory=dao_factory
+                )
+            )
+            file_deletion_requested_dao = await dao.get_file_deletion_requested_dao(
+                dao_factory=dao_factory
+            )
+            idempotence_handler = IdempotenceHandler(
+                file_registry=file_registry,
+                nonstaged_file_requested_dao=nonstaged_file_requested_dao,
+                file_upload_validation_success_dao=file_upload_validation_success_dao,
+                file_deletion_requested_dao=file_deletion_requested_dao,
+            )
+
         outbox_translators = [
-            FileDeletionRequestedListener(config=config, file_registry=file_registry),
-            FileValidationSuccessListener(config=config, file_registry=file_registry),
-            NonstagedFileRequestedListener(config=config, file_registry=file_registry),
+            cls(config=config, idempotence_handler=idempotence_handler)
+            for cls in (
+                FileDeletionRequestedListener,
+                FileValidationSuccessListener,
+                NonstagedFileRequestedListener,
+            )
         ]
         async with KafkaOutboxSubscriber.construct(
             config=config, translators=outbox_translators
